+++ src/app/modules/animals/data/use-cases/animal-heifer-calf-stages-use-cases/remote-delete-animal-heifer-calf-stage-use-case.ts
