+++ src/app/modules/animals/main/factories/use-cases/animal-heifer-calf-stages-use-cases/remote-get-animal-heifer-calf-stage-additional-data-use-case-factory.ts
@@ -1,6 +1,6 @@
 import { makeApiHttpClient } from '@/core/main/factories/http'
 
-import { RemoteGetAnimalHeiferCalfStageAdditionalDataUseCase } from '../../../../data/use-cases/heifer-calf-stages-use-cases'
+import { RemoteGetAnimalHeiferCalfStageAdditionalDataUseCase } from '../../../../data/use-cases/animal-heifer-calf-stages-use-cases'
 
 import type { AnimalHeiferCalfStageDetailsModel } from '../../../../domain/models/animal-heifer-calf-stages-model'
 import type { GetAnimalHeiferCalfStageAdditionalDataUseCase } from '../../../../domain/use-cases/animal-heifer-calf-stages-use-cases'
