+++ src/app/modules/animals/main/factories/use-cases/animal-heifer-calf-stages-use-cases/remote-get-animal-heifer-calf-stages-use-case-factory.ts
@@ -1,6 +1,6 @@
 import { makeApiHttpClient } from '@/core/main/factories/http'
 
-import { RemoteGetAnimalHeiferCalfStagesUseCase } from '../../../../data/use-cases/heifer-calf-stages-use-cases'
+import { RemoteGetAnimalHeiferCalfStagesUseCase } from '../../../../data/use-cases/animal-heifer-calf-stages-use-cases'
 
 import type { AnimalHeiferCalfStageModel } from '../../../../domain/models/animal-heifer-calf-stages-model'
 import type { GetAnimalHeiferCalfStagesUseCase } from '../../../../domain/use-cases/animal-heifer-calf-stages-use-cases'
